requirements.txt
Pillow==8.3.1
requests==2.26.0
plotly==5.1.0
